Library

QuantumUtils

Modules = [QuantumCollocation.QuantumUtils]

QuantumSystems

QuantumCollocation.QuantumSystems.AbstractQuantumSystem — Type

AbstractQuantumSystem

| -> EmbeddedOperator | | | -> QuantumObjective | | -> Matrix (goal) -

Abstract type for defining systems.

QuantumCollocation.QuantumSystems.QuantumSystem — Type

QuantumSystem <: AbstractQuantumSystem

A struct for storing the isomorphisms of the system's drift and drive Hamiltonians, as well as the system's parameters.

QuantumCollocation.QuantumSystems.QuantumSystem — Method

QuantumSystem(
    H_drift::Matrix{<:Number},
    H_drives::Vector{Matrix{<:Number}};
    params=Dict{Symbol, Any}(),
    kwargs...
)::QuantumSystem

Constructs a QuantumSystem object from the drift and drive Hamiltonian terms.

QuantumCollocation.QuantumSystems.QuantumSystemCoupling — Type

QuantumSystemCoupling <: AbstractQuantumSystem

QuantumCollocation.QuantumSystems.lift — Method

lift(U::AbstractMatrix{<:Number}, qubit_index::Int, n_qubits::Int; levels::Int=size(U, 1))

Lift an operator U acting on a single qubit to an operator acting on the entire system of n_qubits.

QuantumCollocation.QuantumSystems.lift — Method

lift(op::AbstractMatrix{<:Number}, i::Int, subsystem_levels::Vector{Int})

Lift an operator op acting on the i-th subsystem to an operator acting on the entire system with given subsystem levels.

Integrators

QuantumCollocation.Integrators.UnitaryPadeIntegrator — Type