Library

Problem Templates

QuantumCollocation.ProblemTemplates.QuantumStateMinimumTimeProblem — Function

QuantumStateMinimumTimeProblem(traj, sys, obj, integrators, constraints; kwargs...)
QuantumStateMinimumTimeProblem(prob; kwargs...)

Construct a QuantumControlProblem for the minimum time problem of reaching a target state.

Arguments

traj::NamedTrajectory: The initial trajectory.

sys::QuantumSystem: The quantum system.

obj::Objective: The objective function.

integrators::Vector{<:AbstractIntegrator}: The integrators.

constraints::Vector{<:AbstractConstraint}: The constraints.

or

prob::QuantumControlProblem: The quantum control problem.

Keyword Arguments

state_name::Symbol=:ψ̃: The symbol for the state variables.

final_fidelity::Union{Real, Nothing}=nothing: The final fidelity.

D=1.0: The cost weight on the time.

ipopt_options::IpoptOptions=IpoptOptions(): The Ipopt options.

piccolo_options::PiccoloOptions=PiccoloOptions(): The Piccolo options.

kwargs...: Additional keyword arguments, passed to QuantumControlProblem.

QuantumCollocation.ProblemTemplates.QuantumStateSmoothPulseProblem — Function

QuantumStateSmoothPulseProblem(system, ψ_inits, ψ_goals, T, Δt; kwargs...)
QuantumStateSmoothPulseProblem(system, ψ_init, ψ_goal, T, Δt; kwargs...)
QuantumStateSmoothPulseProblem(H_drift, H_drives, args...; kwargs...)

Create a quantum state smooth pulse problem. The goal is to find a control pulse a(t) that drives all of the initial states ψ_inits to the corresponding target states ψ_goals using T timesteps of size Δt. This problem also controls the first and second derivatives of the control pulse, da(t) and dda(t), to ensure smoothness.

Arguments

system::AbstractQuantumSystem: The quantum system.

or

H_drift::AbstractMatrix{<:Number}: The drift Hamiltonian.

H_drives::Vector{<:AbstractMatrix{<:Number}}: The control Hamiltonians.

with

ψ_inits::Vector{<:AbstractVector{<:ComplexF64}}: The initial states.

ψ_goals::Vector{<:AbstractVector{<:ComplexF64}}: The target states.

or

ψ_init::AbstractVector{<:ComplexF64}: The initial state.

ψ_goal::AbstractVector{<:ComplexF64}: The target state.

with

T::Int: The number of timesteps.

Δt::Float64: The timestep size.

Keyword Arguments

ipopt_options::IpoptOptions=IpoptOptions(): The IPOPT options.

piccolo_options::PiccoloOptions=PiccoloOptions(): The Piccolo options.

state_name::Symbol=:ψ̃: The name of the state variable.

control_name::Symbol=:a: The name of the control variable.

timestep_name::Symbol=:Δt: The name of the timestep variable.

init_trajectory::Union{NamedTrajectory, Nothing}=nothing: The initial trajectory.

a_bound::Float64=1.0: The bound on the control pulse.

a_bounds::Vector{Float64}=fill(a_bound, length(system.G_drives)): The bounds on the control pulse.

a_guess::Union{Matrix{Float64}, Nothing}=nothing: The initial guess for the control pulse.

da_bound::Float64=Inf: The bound on the first derivative of the control pulse.

da_bounds::Vector{Float64}=fill(da_bound, length(system.G_drives)): The bounds on the first derivative of the control pulse.

zero_initial_and_final_derivative::Bool=false: Whether to enforce zero initial and final derivative.

dda_bound::Float64=1.0: The bound on the second derivative of the control pulse.

dda_bounds::Vector{Float64}=fill(dda_bound, length(system.G_drives)): The bounds on the second derivative of the control pulse.

Δt_min::Float64=0.5 * Δt: The minimum timestep size.

Δt_max::Float64=1.5 * Δt: The maximum timestep size.

drive_derivative_σ::Float64=0.01: The standard deviation of the drive derivative random initialization.

Q::Float64=100.0: The weight on the state objective.

R=1e-2: The weight on the control pulse and its derivatives.

R_a::Union{Float64, Vector{Float64}}=R: The weight on the control pulse.

R_da::Union{Float64, Vector{Float64}}=R: The weight on the first derivative of the control pulse.

R_dda::Union{Float64, Vector{Float64}}=R: The weight on the second derivative of the control pulse.

leakage_operator::Union{Nothing, EmbeddedOperator}=nothing: The leakage operator, if leakage suppression is desired.

constraints::Vector{<:AbstractConstraint}=AbstractConstraint[]: The constraints.

QuantumCollocation.ProblemTemplates.UnitaryBangBangProblem — Function

UnitaryBangBangProblem(system::QuantumSystem, operator, T, Δt; kwargs...)
UnitaryBangBangProblem(H_drift, H_drives, operator, T, Δt; kwargs...)

Construct a QuantumControlProblem for a free-time unitary gate problem with bang-bang control pulses.

\[\begin{aligned} \underset{\vec{\tilde{U}}, a, \dot{a}, \Delta t}{\text{minimize}} & \quad Q \cdot \ell\qty(\vec{\tilde{U}}_T, \vec{\tilde{U}}_{\text{goal}}) + R_{\text{bang-bang}} \cdot \sum_t |\dot{a}_t| \\ \text{ subject to } & \quad \vb{P}^{(n)}\qty(\vec{\tilde{U}}_{t+1}, \vec{\tilde{U}}_t, a_t, \Delta t_t) = 0 \\ & \quad a_{t+1} - a_t - \dot{a}_t \Delta t_t = 0 \\ & \quad |a_t| \leq a_{\text{bound}} \\ & \quad |\dot{a}_t| \leq da_{\text{bound}} \\ & \quad \Delta t_{\text{min}} \leq \Delta t_t \leq \Delta t_{\text{max}} \\ \end{aligned}\]

where, for $U \in SU(N)$,

\[\ell\qty(\vec{\tilde{U}}_T, \vec{\tilde{U}}_{\text{goal}}) = \abs{1 - \frac{1}{N} \abs{ \tr \qty(U_{\text{goal}}, U_T)} }\]

is the infidelity objective function, $Q$ is a weight, $R_a$, and $R_{\dot{a}}$ are weights on the regularization terms, and $\vb{P}^{(n)}$ is the $n$th-order Pade integrator.

TODO: Document bang-bang modification.

Arguments

H_drift::AbstractMatrix{<:Number}: the drift hamiltonian

H_drives::Vector{<:AbstractMatrix{<:Number}}: the control hamiltonians

or

system::QuantumSystem: the system to be controlled

with

operator::AbstractPiccoloOperator: the target unitary, either in the form of an EmbeddedOperator or a `Matrix{ComplexF64}

T::Int: the number of timesteps

Δt::Float64: the (initial) time step size

Keyword Arguments

ipopt_options::IpoptOptions=IpoptOptions(): the options for the Ipopt solver

piccolo_options::PiccoloOptions=PiccoloOptions(): the options for the Piccolo solver

state_name::Symbol = :Ũ⃗: the name of the state variable

control_name::Symbol = :a: the name of the control variable

timestep_name::Symbol = :Δt: the name of the timestep variable

init_trajectory::Union{NamedTrajectory, Nothing}=nothing: an initial trajectory to use

a_bound::Float64=1.0: the bound on the control pulse

a_bounds=fill(a_bound, length(system.G_drives)): the bounds on the control pulses, one for each drive

a_guess::Union{Matrix{Float64}, Nothing}=nothing: an initial guess for the control pulses

da_bound::Float64=1.0: the bound on the control pulse derivative

da_bounds=fill(da_bound, length(system.G_drives)): the bounds on the control pulse derivatives, one for each drive

Δt_min::Float64=Δt isa Float64 ? 0.5 * Δt : 0.5 * mean(Δt): the minimum time step size

Δt_max::Float64=Δt isa Float64 ? 1.5 * Δt : 1.5 * mean(Δt): the maximum time step size

drive_derivative_σ::Float64=0.01: the standard deviation of the initial guess for the control pulse derivatives

Q::Float64=100.0: the weight on the infidelity objective

R=1e-2: the weight on the regularization terms

quadratic_control_regularization=false: whether or not to use quadratic regularization for the control pulses

R_a::Union{Float64, Vector{Float64}}=R: the weight on the regularization term for the control pulses

R_da::Union{Float64, Vector{Float64}}=R: the weight on the regularization term for the control pulse derivatives

R_bang_bang::Union{Float64, Vector{Float64}}=1e-1: the weight on the bang-bang regularization term

phase_operators::Union{AbstractVector{<:AbstractMatrix}, Nothing}=nothing: the phase operators for free phase corrections

constraints::Vector{<:AbstractConstraint}=AbstractConstraint[]: the constraints to enforce

QuantumCollocation.ProblemTemplates.UnitaryDirectSumProblem — Method

UnitaryDirectSumProblem(probs, final_fidelity; kwargs...)

Construct a QuantumControlProblem as a direct sum of unitary gate problems. The purpose is to find solutions that are as close as possible with respect to one of their components. In particular, this is useful for finding interpolatable control solutions.

A graph of edges (specified by problem labels) will enforce a PairwiseQuadraticRegularizer between the component trajectories of the problem in probs corresponding to the names of the edge in edges with corresponding edge weight Q.

Boundary values can be included to enforce a QuadraticRegularizer on edges where one of the nodes is not optimized. The boundary values are specified as a dictionary with keys corresponding to the edge labels and values corresponding to the boundary values.

The default behavior is to use a 1D chain for the graph, i.e., enforce a PairwiseQuadraticRegularizer between each neighbor of the provided probs.

Arguments

probs::AbstractVector{<:QuantumControlProblem}: the problems to combine

final_fidelity::Real: the fidelity to enforce between the component final unitaries and the component goal unitaries

Keyword Arguments

prob_labels::AbstractVector{<:String}}: the labels for the problems

graph::Union{Nothing, AbstractVector{<:Tuple{String, String}}, AbstractVector{<:Tuple{Symbol, Symbol}}}`: the graph of edges to enforce

boundary_values::Union{Nothing, AbstractDict{<:String, <:AbstractArray}, AbstractDict{<:Symbol, <:AbstractVector}}=nothing: the boundary values for the problems

Q::Union{Float64, Vector{Float64}}=100.0: the weights on the pairwise regularizers

Q_symb::Symbol=:Ũ⃗: the symbol to use for the regularizer

R::Float64=1e-2: the shared weight on all control terms (:a, :da, :dda is assumed)

R_a::Union{Float64, Vector{Float64}}=R: the weight on the regularization term for the control pulses

R_da::Union{Float64, Vector{Float64}}=R: the weight on the regularization term for the control pulse derivatives

R_dda::Union{Float64, Vector{Float64}}=R: the weight on the regularization term for the control pulse second derivatives

R_b::Union{Float64, Vector{Float64}}=R: the weight on the regularization term for the boundary values

drive_derivative_σ::Float64=0.01: the standard deviation of the initial guess for the control pulse derivatives

drive_reset_ratio::Float64=0.1: amount of random noise to add to the control data (can help avoid hitting restoration if provided problems are converged)

fidelity_cost::Bool=false: whether or not to include a fidelity cost in the objective

subspace::Union{AbstractVector{<:Integer}, Nothing}=nothing: the subspace to use for the fidelity of each problem

ipopt_options::IpoptOptions=IpoptOptions(): the options for the Ipopt solver

piccolo_options::PiccoloOptions=PiccoloOptions(): the options for the Piccolo solver

QuantumCollocation.ProblemTemplates.UnitaryMinimumTimeProblem — Function

UnitaryMinimumTimeProblem(
    trajectory::NamedTrajectory,
    system::AbstractQuantumSystem,
    objective::Objective,
    integrators::Vector{<:AbstractIntegrator},
    constraints::Vector{<:AbstractConstraint};
    kwargs...
)

UnitaryMinimumTimeProblem(
    prob::QuantumControlProblem;
    kwargs...
)

Create a minimum-time problem for unitary control.

\[\begin{aligned} \underset{\vec{\tilde{U}}, a, \dot{a}, \ddot{a}, \Delta t}{\text{minimize}} & \quad J(\vec{\tilde{U}}, a, \dot{a}, \ddot{a}) + D \sum_t \Delta t_t \\ \text{ subject to } & \quad \vb{P}^{(n)}\qty(\vec{\tilde{U}}_{t+1}, \vec{\tilde{U}}_t, a_t, \Delta t_t) = 0 \\ & c(\vec{\tilde{U}}, a, \dot{a}, \ddot{a}) = 0 \\ & \quad \Delta t_{\text{min}} \leq \Delta t_t \leq \Delta t_{\text{max}} \\ \end{aligned}\]

Arguments

trajectory::NamedTrajectory: The initial trajectory.

system::AbstractQuantumSystem: The quantum system.

objective::Objective: The objective function (additional to the minimum-time objective).

integrators::Vector{<:AbstractIntegrator}: The integrators.

constraints::Vector{<:AbstractConstraint}: The constraints.

Keyword Arguments

unitary_name::Symbol=:Ũ⃗: The symbol for the unitary control.

final_fidelity::Float64=0.99: The final fidelity.

D=1.0: The weight for the minimum-time objective.

ipopt_options::IpoptOptions=IpoptOptions(): The options for the Ipopt solver.

piccolo_options::PiccoloOptions=PiccoloOptions(): The options for the Piccolo solver.

kwargs...: Additional keyword arguments to pass to QuantumControlProblem.

QuantumCollocation.ProblemTemplates.UnitaryRobustnessProblem — Function

UnitaryRobustnessProblem(
    H_error,
    trajectory,
    system,
    objective,
    integrators,
    constraints;
    kwargs...
)

UnitaryRobustnessProblem(Hₑ, prob::QuantumControlProblem; kwargs...)

Create a quantum control problem for robustness optimization of a unitary trajectory.

Keyword Arguments

unitary_name::Symbol=:Ũ⃗: The symbol for the unitary trajectory in trajectory.

final_fidelity::Union{Real, Nothing}=nothing: The target fidelity for the final unitary.

ipopt_options::IpoptOptions=IpoptOptions(): Options for the Ipopt solver.

piccolo_options::PiccoloOptions=PiccoloOptions(): Options for the Piccolo solver.

kwargs...: Additional keyword arguments passed to QuantumControlProblem.

QuantumCollocation.ProblemTemplates.UnitarySamplingProblem — Method

UnitarySamplingProblem(systemns, operator, T, Δt; kwargs...)

A UnitarySamplingProblem is a quantum control problem where the goal is to find a control pulse that generates a target unitary operator for a set of quantum systems. The controls are shared among all systems, and the optimization seeks to find the control pulse that achieves the operator for each system. The idea is to enforce a robust solution by including multiple systems reflecting the problem uncertainty.

Arguments

systems::AbstractVector{<:AbstractQuantumSystem}: A vector of quantum systems.

operator::AbstractPiccoloOperator: The target unitary operator.

T::Int: The number of time steps.

Δt::Union{Float64, Vector{Float64}}: The time step value or vector of time steps.

Keyword Arguments

system_labels::Vector{String} = string.(1:length(systems)): The labels for each system.

system_weights::Vector{Float64} = fill(1.0, length(systems)): The weights for each system.

init_trajectory::Union{NamedTrajectory, Nothing} = nothing: The initial trajectory.

ipopt_options::IpoptOptions = IpoptOptions(): The IPOPT options.

piccolo_options::PiccoloOptions = PiccoloOptions(): The Piccolo options.

state_name::Symbol = :Ũ⃗: The name of the state variable.

control_name::Symbol = :a: The name of the control variable.

timestep_name::Symbol = :Δt: The name of the timestep variable.

constraints::Vector{<:AbstractConstraint} = AbstractConstraint[]: The constraints.

a_bound::Float64 = 1.0: The bound for the control amplitudes.

a_bounds::Vector{Float64} = fill(a_bound, length(systems[1].G_drives)): The bounds for the control amplitudes.

a_guess::Union{Matrix{Float64}, Nothing} = nothing: The initial guess for the control amplitudes.

da_bound::Float64 = Inf: The bound for the control first derivatives.

da_bounds::Vector{Float64} = fill(da_bound, length(systems[1].G_drives)): The bounds for the control first derivatives.

dda_bound::Float64 = 1.0: The bound for the control second derivatives.

dda_bounds::Vector{Float64} = fill(dda_bound, length(systems[1].G_drives)): The bounds for the control second derivatives.

Δt_min::Float64 = 0.5 * Δt: The minimum time step size.

Δt_max::Float64 = 1.5 * Δt: The maximum time step size.

Q::Float64 = 100.0: The fidelity weight.

R::Float64 = 1e-2: The regularization weight.

R_a::Union{Float64, Vector{Float64}} = R: The regularization weight for the control amplitudes.

R_da::Union{Float64, Vector{Float64}} = R: The regularization weight for the control first derivatives.

R_dda::Union{Float64, Vector{Float64}} = R: The regularization weight for the control second derivatives.

kwargs...: Additional keyword arguments.

QuantumCollocation.ProblemTemplates.UnitarySmoothPulseProblem — Method

UnitarySmoothPulseProblem(system::AbstractQuantumSystem, operator, T, Δt; kwargs...)
UnitarySmoothPulseProblem(H_drift, H_drives, operator, T, Δt; kwargs...)

Construct a QuantumControlProblem for a free-time unitary gate problem with smooth control pulses enforced by constraining the second derivative of the pulse trajectory, i.e.,

\[\begin{aligned} \underset{\vec{\tilde{U}}, a, \dot{a}, \ddot{a}, \Delta t}{\text{minimize}} & \quad Q \cdot \ell\qty(\vec{\tilde{U}}_T, \vec{\tilde{U}}_{\text{goal}}) + \frac{1}{2} \sum_t \qty(R_a a_t^2 + R_{\dot{a}} \dot{a}_t^2 + R_{\ddot{a}} \ddot{a}_t^2) \\ \text{ subject to } & \quad \vb{P}^{(n)}\qty(\vec{\tilde{U}}_{t+1}, \vec{\tilde{U}}_t, a_t, \Delta t_t) = 0 \\ & \quad a_{t+1} - a_t - \dot{a}_t \Delta t_t = 0 \\ & \quad \dot{a}_{t+1} - \dot{a}_t - \ddot{a}_t \Delta t_t = 0 \\ & \quad |a_t| \leq a_{\text{bound}} \\ & \quad |\ddot{a}_t| \leq \ddot{a}_{\text{bound}} \\ & \quad \Delta t_{\text{min}} \leq \Delta t_t \leq \Delta t_{\text{max}} \\ \end{aligned}\]

where, for $U \in SU(N)$,

\[\ell\qty(\vec{\tilde{U}}_T, \vec{\tilde{U}}_{\text{goal}}) = \abs{1 - \frac{1}{N} \abs{ \tr \qty(U_{\text{goal}}, U_T)} }\]

is the infidelity objective function, $Q$ is a weight, $R_a$, $R_{\dot{a}}$, and $R_{\ddot{a}}$ are weights on the regularization terms, and $\vb{P}^{(n)}$ is the $n$th-order Pade integrator.

Arguments

system::AbstractQuantumSystem: the system to be controlled

or

H_drift::AbstractMatrix{<:Number}: the drift hamiltonian

H_drives::Vector{<:AbstractMatrix{<:Number}}: the control hamiltonians

with

operator::AbstractPiccoloOperator: the target unitary, either in the form of an EmbeddedOperator or a `Matrix{ComplexF64}

T::Int: the number of timesteps

Δt::Float64: the (initial) time step size

Keyword Arguments

ipopt_options::IpoptOptions=IpoptOptions(): the options for the Ipopt solver

piccolo_options::PiccoloOptions=PiccoloOptions(): the options for the Piccolo solver

state_name::Symbol = :Ũ⃗: the name of the state

control_name::Symbol = :a: the name of the control

timestep_name::Symbol = :Δt: the name of the timestep

init_trajectory::Union{NamedTrajectory, Nothing}=nothing: an initial trajectory to use

a_bound::Float64=1.0: the bound on the control pulse

a_bounds::Vector{Float64}=fill(a_bound, length(system.G_drives)): the bounds on the control pulses, one for each drive

a_guess::Union{Matrix{Float64}, Nothing}=nothing: an initial guess for the control pulses

da_bound::Float64=Inf: the bound on the control pulse derivative

da_bounds::Vector{Float64}=fill(da_bound, length(system.G_drives)): the bounds on the control pulse derivatives, one for each drive

zero_initial_and_final_derivative::Bool=false: whether to enforce zero initial and final control pulse derivatives

dda_bound::Float64=1.0: the bound on the control pulse second derivative

dda_bounds::Vector{Float64}=fill(dda_bound, length(system.G_drives)): the bounds on the control pulse second derivatives, one for each drive

Δt_min::Float64=Δt isa Float64 ? 0.5 * Δt : 0.5 * mean(Δt): the minimum time step size

Δt_max::Float64=Δt isa Float64 ? 1.5 * Δt : 1.5 * mean(Δt): the maximum time step size

Q::Float64=100.0: the weight on the infidelity objective

R=1e-2: the weight on the regularization terms

R_a::Union{Float64, Vector{Float64}}=R: the weight on the regularization term for the control pulses

R_da::Union{Float64, Vector{Float64}}=R: the weight on the regularization term for the control pulse derivatives

R_dda::Union{Float64, Vector{Float64}}=R: the weight on the regularization term for the control pulse second derivatives

phase_name::Symbol=:ϕ: the name of the phase

phase_operators::Union{AbstractVector{<:AbstractMatrix}, Nothing}=nothing: the phase operators for free phase corrections

constraints::Vector{<:AbstractConstraint}=AbstractConstraint[]: the constraints to enforce

Options

QuantumCollocationCore.Options.IpoptOptions — Type

Solver options for Ipopt

https://coin-or.github.io/Ipopt/OPTIONS.html#OPT_print_options_documentation

QuantumCollocationCore.Options.PiccoloOptions — Type

PiccoloOptions

Options for the Piccolo quantum optimal control library.

Fields

verbose::Bool = true: Print verbose output

verbose_evaluator::Bool = false: Print verbose output from the evaluator

free_time::Bool = true: Allow free time optimization

timesteps_all_equal::Bool = true: Use equal timesteps

integrator::Symbol = :pade: Integrator to use

pade_order::Int = 4: Order of the Pade approximation

rollout_integrator::Function = expv: Integrator to use for rollout

jacobian_structure::Bool = integrator == :pade: Use the integrator's Jacobian structure

eval_hessian::Bool = false: Evaluate the Hessian

geodesic = true: Use the geodesic to initialize the optimization.

blas_multithreading::Bool = true: Use BLAS multithreading.

build_trajectory_constraints::Bool = true: Build trajectory constraints.

complex_control_norm_constraint_name::Union{Nothing, Symbol} = nothing: Name of the complex control norm constraint.

complex_control_norm_constraint_radius::Float64 = 1.0: Radius of the complex control norm constraint.

bound_state::Bool = false: Bound the state.

leakage_suppression::Bool = false: Suppress leakage.

R_leakage::Float64 = 1.0: Leakage suppression parameter.

free_phase_infidelity::Bool = false: Free phase infidelity.

phase_operators::Union{Nothing, AbstractVector{<:AbstractMatrix{<:Complex}}} = nothing: Phase operators.

phase_name::Symbol = :ϕ: Name of the phase.

Rollouts

QuantumCollocation.Rollouts.infer_is_evp — Method

infer_is_evp(integrator::Function)

Infer whether the integrator is a exponential-vector product (EVP) function.

If true, the integrator is expected to have a signature like the exponential action, expv. Otherwise, it is expected to have a signature like exp.

QuantumCollocation.Rollouts.lab_frame_unitary_rollout — Method

lab_frame_unitary_rollout(
    sys::AbstractQuantumSystem,
    controls::AbstractMatrix{Float64};
    duration=nothing,
    timestep=nothing,
    ω=nothing,
    timestep_nyquist=1 / (50 * ω)
)

QuantumCollocation.Rollouts.rollout — Method

rollout(
    ψ̃_init::AbstractVector{<:Float64},
    controls::AbstractMatrix,
    Δt::AbstractVector,
    system::AbstractQuantumSystem
)

Rollout a quantum state ψ̃_init under the control controls for a time Δt using the system system.

If exp_vector_product is true, the integrator is expected to have a signature like the exponential action, expv. Otherwise, it is expected to have a signature like exp.

Types should allow for autodifferentiable controls and times.

Direct Sums

QuantumCollocation.DirectSums.direct_sum — Method

direct_sum(A::AbstractMatrix, B::AbstractMatrix)

Returns the direct sum of two matrices.

QuantumCollocation.DirectSums.direct_sum — Method

direct_sum(Ã⃗::AbstractVector, B̃⃗::AbstractVector)

Returns the direct sum of two iso_vec operators.

QuantumCollocation.DirectSums.direct_sum — Method

direct_sum(sys1::QuantumSystem, sys2::QuantumSystem)

Returns the direct sum of two QuantumSystem objects.

QuantumCollocation.DirectSums.direct_sum — Method

direct_sum(A::SparseMatrixCSC, B::SparseMatrixCSC)

Returns the direct sum of two sparse matrices.

Trajectory Initialization

QuantumCollocation.TrajectoryInitialization.initialize_trajectory — Method

initialize_trajectory

Trajectory initialization of unitaries.

QuantumCollocation.TrajectoryInitialization.initialize_trajectory — Method

initialize_trajectory

Trajectory initialization of quantum states.

QuantumCollocation.TrajectoryInitialization.initialize_trajectory — Method

initialize_trajectory

Initialize a trajectory for a control problem. The trajectory is initialized with data that should be consistently the same type (in this case, Float64).

QuantumCollocation.TrajectoryInitialization.unitary_geodesic — Function

unitary_geodesic(
    operator::EmbeddedOperator,
    samples::Int;
    kwargs...
)

unitary_geodesic(
    U_goal::AbstractMatrix{<:Number},
    samples::Int;
    kwargs...
)

unitary_geodesic(
    U₀::AbstractMatrix{<:Number},
    U₁::AbstractMatrix{<:Number},
    samples::Number;
    kwargs...
)

unitary_geodesic(
    U₀::AbstractMatrix{<:Number},
    U₁::AbstractMatrix{<:Number},
    timesteps::AbstractVector{<:Number};
    return_generator=false
)

Compute a geodesic connecting two unitary operators.

QuantumCollocation.TrajectoryInitialization.unitary_geodesic — Method

unitary_geodesic(U_init, U_goal, times; kwargs...)

Compute the geodesic connecting Uinit and Ugoal at the specified times. Allows for the possibility of unequal times and ranges outside [0,1].

Arguments

U_init::AbstractMatrix{<:Number}: The initial unitary operator.

U_goal::AbstractMatrix{<:Number}: The goal unitary operator.

times::AbstractVector{<:Number}: The times at which to evaluate the geodesic.

Keyword Arguments

return_unitary_isos::Bool=true: If true returns a matrix where each column is a unitary isovec, i.e. vec(vcat(real(U), imag(U))). If false, returns a vector of unitary matrices.

return_generator::Bool=false: If true, returns the effective Hamiltonian generating the geodesic.

QuantumCollocation.TrajectoryInitialization.unitary_linear_interpolation — Method

unitary_linear_interpolation(
    U_init::AbstractMatrix,
    U_goal::AbstractMatrix,
    samples::Int
)

Compute a linear interpolation of unitary operators with samples samples.

Trajectory Interpolations