Library

QuantumUtils

QuantumCollocation.QuantumUtils.GATES — Constant

GATES::Dict{Symbol, Matrix{ComplexF64}}

Dictionary of common quantum gates.

Elements

:I - identity, $I$

:X - Pauli X, $\sigma_x$

:Y - Pauli Y, $\sigma_y$

:Z - Pauli Z, $\sigma_z$

:H - Hadamard, $H$

:CX - controlled-X, $C_X$

:XI - $-i X \otimes I$

:sqrtiSWAP - $\sqrt{i \text{SWAP}}$

QuantumCollocation.QuantumUtils.:⊗ — Method

⊗(A::AbstractVecOrMat, B::AbstractVecOrMat)

Kronecker product of two vectors or matrices.

QuantumCollocation.QuantumUtils.apply — Method

apply(gate::Symbol, ψ::Vector{<:Number})

Apply a gate from the GATES dictionary to a quantum state.

QuantumCollocation.QuantumUtils.gate — Method

gate(U::Symbol)

Get a gate from the GATES dictionary.

QuantumCollocation.QuantumUtils.lift — Method

lift(
    U::AbstractMatrix{<:Number},
    i::Int,
    n::Int;
    levels::Int=size(U, 1)
)

Lift a matrix to a larger Hilbert space by placing it on the i-th qubit of a n-qubit system. I.e.,

\[U \mapsto I_1 \otimes \cdots \otimes I_{i-1} \otimes U \otimes I_{i+1} \otimes \cdots \otimes I_n\]

QuantumCollocation.QuantumUtils.lift — Method

lift(
    op::AbstractMatrix{<:Number},
    i::Int,
    levels::Vector{Int}
)

Lift a matrix to a larger Hilbert space by placing it on the i-th qubit of a n-qubit system. I.e.,

\[U \mapsto I_1 \otimes \cdots \otimes I_{i-1} \otimes U \otimes I_{i+1} \otimes \cdots \otimes I_n\]

QuantumCollocation.QuantumUtils.qubit_system_state — Method

qubit_system_state(ket::String)

Get a quantum state from a string of qubit states, e.g. "01" -> $\ket{01}$.

QuantumSystems

QuantumCollocation.QuantumSystems.AbstractSystem — Type

AbstractSystem

Abstract type for defining systems.

QuantumCollocation.QuantumSystems.QuantumSystem — Type

QuantumSystemNew <: AbstractSystem

A struct for storing the isomorphisms of the system's drift and drive Hamiltonians, as well as the system's parameters.

QuantumCollocation.QuantumSystems.QuantumSystem — Method

QuantumSystem(
    H_drift::Matrix{<:Number},
    H_drives::Vector{Matrix{<:Number}};
    params=Dict{Symbol, Any}(),
    kwargs...
)::QuantumSystem

Constructs a QuantumSystem object from the drift and drive Hamiltonian terms.

QuantumCollocation.QuantumSystems.G — Method

G(H::AbstractMatrix)::Matrix{Float64}

Returns the isomorphism of $-iH$:

\[G(H) = \widetilde{- i H} = \mqty(1 & 0 \\ 0 & 1) \otimes \Im(H) - \mqty(0 & -1 \\ 1 & 0) \otimes \Re(H)\]

where $\Im(H)$ and $\Re(H)$ are the imaginary and real parts of $H$ and the tilde indicates the standard isomorphism of a complex valued matrix:

\[\widetilde{H} = \mqty(1 & 0 \\ 0 & 1) \otimes \Re(H) + \mqty(0 & -1 \\ 1 & 0) \otimes \Im(H)\]

QuantumCollocation.QuantumSystems.H — Method

H(G::AbstractMatrix{<:Number})::Matrix{ComplexF64}

Returns the inverse of G(H) = iso(-iH), i.e. returns H

QuantumCollocation.QuantumSystems.MultiModeSystem — Method

MultiModeSystem(
    transmon_levels::Int,
    cavity_levels::Int;
    χ=2π * -0.5459e-3,
    κ=2π * 4e-6,
    χGF=2π * -1.01540302914e-3,
    α=-2π * 0.143,
    n_cavities=1
)::QuantumSystem

Create a new QuantumSystemNew object for a transmon qubit with transmon_levels levels coupled to a single cavity with cavity_levels levels.

The Hamiltonian for this system is given by

\[\hat H = \frac{\kappa}{2} \hat{a}^{ \dagger} \hat{a} \left( \hat{a}^{\dagger}\hat{a}-1\right) + 2 \chi \dyad{e} \hat{a}^{\dagger}\hat{a} + \left( \epsilon_{c}(t) + \epsilon_{q}(t) + \mathrm{c.c.} \right)\]

QuantumCollocation.QuantumSystems.lie_subalgebra_dim — Method

lie_subalgebra_dim(Hs::Vector{<:AbstractMatrix})

Returns the dimension of the Lie subalgebra generated by the operators in Hs.

Integrators

QuantumCollocation.Integrators.UnitaryPadeIntegrator — Type