Library

QuantumUtils

QuantumCollocation.QuantumUtils.GATES — Constant

A constant dictionary GATES containing common quantum gate matrices as complex-valued matrices. Each gate is represented by its unitary matrix.

GATES[:I] - Identity gate: Leaves the state unchanged.

GATES[:X] - Pauli-X (NOT) gate: Flips the qubit state.

GATES[:Y] - Pauli-Y gate: Rotates the qubit state around the Y-axis of the Bloch sphere.

GATES[:Z] - Pauli-Z gate: Flips the phase of the qubit state.

GATES[:H] - Hadamard gate: Creates superposition by transforming basis states.

GATES[:CX] - Controlled-X (CNOT) gate: Flips the second qubit (target) if the first qubit (control) is |1⟩.

GATES[:CZ] - Controlled-Z (CZ) gate: Flips the phase of the second qubit (target) if the first qubit (control) is |1⟩.

GATES[:XI] - Complex gate: A specific gate used for complex operations.

GATES[:sqrtiSWAP] - Square root of iSWAP gate: Partially swaps two qubits with a phase.

julia> GATES[:Z]
2×2 Matrix{ComplexF64}:
 1.0+0.0im   0.0+0.0im
 0.0+0.0im  -1.0+0.0im

julia> get_gate(:CX)
4×4 Matrix{ComplexF64}:
 1.0+0.0im  0.0+0.0im  0.0+0.0im  0.0+0.0im
 0.0+0.0im  1.0+0.0im  0.0+0.0im  0.0+0.0im
 0.0+0.0im  0.0+0.0im  0.0+0.0im  1.0+0.0im
 0.0+0.0im  0.0+0.0im  1.0+0.0im  0.0+0.0im

QuantumCollocation.QuantumUtils.:⊗ — Method

⊗(A::AbstractVecOrMat, B::AbstractVecOrMat) = kron(A, B)

The Kronecker product, denoted by ⊗, results in a block matrix formed by multiplying each element of A by the entire matrix B.

julia> GATES[:X] ⊗ GATES[:Y]
4×4 Matrix{ComplexF64}:
 0.0+0.0im  0.0+0.0im  0.0+0.0im  0.0-1.0im
 0.0+0.0im  0.0+0.0im  0.0+1.0im  0.0+0.0im
 0.0+0.0im  0.0-1.0im  0.0+0.0im  0.0+0.0im
 0.0+1.0im  0.0+0.0im  0.0+0.0im  0.0+0.0im

QuantumCollocation.QuantumUtils.annihilate — Method

annihilate(levels::Int)

Get the annihilation operator for a system with levels levels.

QuantumCollocation.QuantumUtils.apply — Method

apply(gate::Symbol, ψ::Vector{<:Number})

Apply a quantum gate gate to a state vector ψ.

QuantumCollocation.QuantumUtils.cavity_state — Method

cavity_state(level::Int, cavity_levels::Int)

Generate the state vector for a given level in a cavity system with cavity_levels levels.

Arguments

level::Int: The index of the desired level (must be between 0 and cavity_levels - 1).

cavity_levels::Int: The total number of levels in the cavity system.

Returns

Vector{ComplexF64}: A state vector with cavity_levels elements where the element at level + 1 is 1.0 and all other elements are 0.0.

Throws

BoundsError: If level is less than 0 or greater than or equal to cavity_levels.

Examples

julia> cavity_state(2, 5)
5-element Vector{ComplexF64}:
 0.0 + 0.0im
 0.0 + 0.0im
 1.0 + 0.0im
 0.0 + 0.0im
 0.0 + 0.0im

QuantumCollocation.QuantumUtils.create — Method

create(levels::Int)

Get the creation operator for a system with levels levels.

QuantumCollocation.QuantumUtils.fidelity — Method

fidelity(ψ, ψ_goal; subspace=1:length(ψ))

Calculate the fidelity between two quantum states ψ and ψ_goal.

QuantumCollocation.QuantumUtils.haar_identity — Method

haar_identity(n::Int, radius::Number)

Generate a random unitary matrix close to the identity matrix using the Haar measure for an n-dimensional system with a given radius.

QuantumCollocation.QuantumUtils.haar_random — Method

haar_random(n::Int)

Generate a random unitary matrix using the Haar measure for an n-dimensional system.

QuantumCollocation.QuantumUtils.iso_fidelity — Method

iso_fidelity(ψ̃, ψ̃_goal; kwargs...)

Calculate the fidelity between two quantum states in their isomorphic form ψ̃ and ψ̃_goal.

QuantumCollocation.QuantumUtils.iso_operator_to_iso_vec — Method

iso_operator_to_iso_vec(Ũ::AbstractMatrix{R}) where R <: Real

Convert a real matrix Ũ representing an isomorphism operator into a real vector.

QuantumCollocation.QuantumUtils.iso_to_ket — Method

iso_to_ket(ψ̃)

Convert a complex vector ψ̃ with real and imaginary parts into a ket vector.

QuantumCollocation.QuantumUtils.iso_vec_to_iso_operator — Method

iso_vec_to_iso_operator(Ũ⃗::AbstractVector{R}) where R <: Real

Convert a real vector Ũ⃗ into a real matrix representing an isomorphism operator.

QuantumCollocation.QuantumUtils.iso_vec_to_operator — Method

iso_vec_to_operator(Ũ⃗::AbstractVector{R}) where R <: Real

Convert a real vector Ũ⃗ into a complex matrix representing an operator.

QuantumCollocation.QuantumUtils.ket_to_iso — Method

ket_to_iso(ψ)

Convert a ket vector ψ into a complex vector with real and imaginary parts.

QuantumCollocation.QuantumUtils.kron_from_dict — Method

kronfromdict(keys, dict; kwargs...)

Reduce the keys to a single operator by using the provided dictionary and the kronecker product.

QuantumCollocation.QuantumUtils.lift — Method

lift(U::AbstractMatrix{<:Number}, qubit_index::Int, n_qubits::Int; levels::Int=size(U, 1))

Lift an operator U acting on a single qubit to an operator acting on the entire system of n_qubits.

QuantumCollocation.QuantumUtils.lift — Method

lift(op::AbstractMatrix{<:Number}, i::Int, subsystem_levels::Vector{Int})

Lift an operator op acting on the i-th subsystem to an operator acting on the entire system with given subsystem levels.

QuantumCollocation.QuantumUtils.number — Method

number(levels::Int)

Get the number operator n = a'a for a system with levels levels.

QuantumCollocation.QuantumUtils.operator_to_iso_vec — Method

operator_to_iso_vec(U::AbstractMatrix{<:Complex})

Convert a complex matrix U representing an operator into a real vector.

QuantumCollocation.QuantumUtils.operators_from_dict — Method

operatorsfromdict(keys::AbstractVector{<:Any}, operatordictionary; Ikey=:I)

Replace the vector of keys using the operators from a dictionary.

QuantumCollocation.QuantumUtils.operators_from_dict — Method

operatorsfromdict(keystring::String, operatordictionary; I_key="I")

Replace the string (each character is one key) with operators from a dictionary.

QuantumCollocation.QuantumUtils.population — Method

population(ψ̃, i)

Calculate the population of the i-th level for a given state vector ψ̃ in its isomorphic form.

QuantumCollocation.QuantumUtils.populations — Method

populations(ψ::AbstractVector{<:Complex})

Calculate the populations for each level of a given state vector ψ.

QuantumCollocation.QuantumUtils.populations — Method

populations(ψ̃::AbstractVector{<:Real})

Calculate the populations for each level of a given state vector ψ̃ in its isomorphic form.

QuantumCollocation.QuantumUtils.quad — Method

quad(levels::Int)

Get the operator n(n - I) for a system with levels levels.

QuantumCollocation.QuantumUtils.quantum_state — Method

quantum_state(
    ket::String,
    levels::Vector{Int};
    level_dict=Dict(:g => 0, :e => 1, :f => 2, :h => 2),
    return_states=false
)

Construct a quantum state from a string ket representation.

Example

TODO: add example

QuantumCollocation.QuantumUtils.qubit_system_state — Method

qubit_system_state(ket::String)

Get the state vector for a qubit system given a ket string ket of 0s and 1s.

QuantumCollocation.QuantumUtils.unitary_fidelity — Method

unitary_fidelity(Ũ⃗::AbstractVector{<:Real}, Ũ⃗_goal::AbstractVector{<:Real}; subspace=nothing)

Calculate the fidelity between two unitary operators in their isomorphic form Ũ⃗ and Ũ⃗_goal.

QuantumCollocation.QuantumUtils.unitary_fidelity — Method

unitary_fidelity(U::Matrix, U_goal::Matrix; subspace=nothing)

Calculate the fidelity between two unitary operators U and U_goal.

QuantumCollocation.QuantumUtils.vec⁻¹ — Method

vec⁻¹(x::AbstractVector)

Convert a vector x into a square matrix. The length of x must be a perfect square.

QuantumSystems

QuantumCollocation.QuantumSystems.AbstractQuantumSystem — Type

AbstractQuantumSystem

Abstract type for defining systems.

QuantumCollocation.QuantumSystems.QuantumSystem — Type

QuantumSystem <: AbstractQuantumSystem

A struct for storing the isomorphisms of the system's drift and drive Hamiltonians, as well as the system's parameters.

QuantumCollocation.QuantumSystems.QuantumSystem — Method

QuantumSystem(
    H_drift::Matrix{<:Number},
    H_drives::Vector{Matrix{<:Number}};
    params=Dict{Symbol, Any}(),
    kwargs...
)::QuantumSystem

Constructs a QuantumSystem object from the drift and drive Hamiltonian terms.

QuantumCollocation.QuantumSystems.QuantumSystemCoupling — Type

QuantumSystemCoupling <: AbstractQuantumSystem

QuantumCollocation.QuantumSystems.G — Method

G(H::AbstractMatrix)::Matrix{Float64}

Returns the isomorphism of $-iH$:

\[G(H) = \widetilde{- i H} = \mqty(1 & 0 \\ 0 & 1) \otimes \Im(H) - \mqty(0 & -1 \\ 1 & 0) \otimes \Re(H)\]

where $\Im(H)$ and $\Re(H)$ are the imaginary and real parts of $H$ and the tilde indicates the standard isomorphism of a complex valued matrix:

\[\widetilde{H} = \mqty(1 & 0 \\ 0 & 1) \otimes \Re(H) + \mqty(0 & -1 \\ 1 & 0) \otimes \Im(H)\]

QuantumCollocation.QuantumSystems.H — Method

H(G::AbstractMatrix{<:Number})::Matrix{ComplexF64}

Returns the inverse of G(H) = iso(-iH), i.e. returns H

Integrators

QuantumCollocation.Integrators.UnitaryPadeIntegrator — Type