Library

QuantumUtils

QuantumCollocation.QuantumUtils.annihilate — Method

annihilate(levels::Int)

Get the annihilation operator for a system with levels levels.

QuantumCollocation.QuantumUtils.create — Method

create(levels::Int)

Get the creation operator for a system with levels levels.

QuantumCollocation.QuantumUtils.kron_from_dict — Method

kronfromdict(keys, dict; kwargs...)

Reduce the keys to a single operator by using the provided dictionary and the kronecker product.

QuantumCollocation.QuantumUtils.number — Method

number(levels::Int)

Get the number operator n = a'a for a system with levels levels.

QuantumCollocation.QuantumUtils.operators_from_dict — Method

operatorsfromdict(keys::AbstractVector{<:Any}, operatordictionary; Ikey=:I)

Replace the vector of keys using the operators from a dictionary.

QuantumCollocation.QuantumUtils.operators_from_dict — Method

operatorsfromdict(keystring::String, operatordictionary; I_key="I")

Replace the string (each character is one key) with operators from a dictionary.

QuantumCollocation.QuantumUtils.quad — Method

quad(levels::Int)

Get the operator n(n - I) for a system with levels levels.

QuantumCollocation.QuantumUtils.quantum_state — Method

quantum_state(
    ket::String,
    levels::Vector{Int};
    level_dict=Dict(:g => 0, :e => 1, :f => 2, :h => 2),
    return_states=false
)

Construct a quantum state from a string ket representation.

Example

TODO: add example

QuantumCollocation.QuantumUtils.vec⁻¹ — Method

isomporphism utilities

QuantumSystems

QuantumCollocation.QuantumSystems.AbstractQuantumSystem — Type

AbstractQuantumSystem

Abstract type for defining systems.

QuantumCollocation.QuantumSystems.QuantumSystem — Type

QuantumSystem <: AbstractQuantumSystem

A struct for storing the isomorphisms of the system's drift and drive Hamiltonians, as well as the system's parameters.

QuantumCollocation.QuantumSystems.QuantumSystem — Method

QuantumSystem(
    H_drift::Matrix{<:Number},
    H_drives::Vector{Matrix{<:Number}};
    params=Dict{Symbol, Any}(),
    kwargs...
)::QuantumSystem

Constructs a QuantumSystem object from the drift and drive Hamiltonian terms.

QuantumCollocation.QuantumSystems.QuantumSystemCoupling — Type

QuantumSystemCoupling <: AbstractQuantumSystem

QuantumCollocation.QuantumSystems.G — Method

G(H::AbstractMatrix)::Matrix{Float64}

Returns the isomorphism of $-iH$:

\[G(H) = \widetilde{- i H} = \mqty(1 & 0 \\ 0 & 1) \otimes \Im(H) - \mqty(0 & -1 \\ 1 & 0) \otimes \Re(H)\]

where $\Im(H)$ and $\Re(H)$ are the imaginary and real parts of $H$ and the tilde indicates the standard isomorphism of a complex valued matrix:

\[\widetilde{H} = \mqty(1 & 0 \\ 0 & 1) \otimes \Re(H) + \mqty(0 & -1 \\ 1 & 0) \otimes \Im(H)\]

QuantumCollocation.QuantumSystems.H — Method

H(G::AbstractMatrix{<:Number})::Matrix{ComplexF64}

Returns the inverse of G(H) = iso(-iH), i.e. returns H

Integrators

QuantumCollocation.Integrators.UnitaryPadeIntegrator — Type