Library

Problem Templates

QuantumCollocation.ProblemTemplates.QuantumStateMinimumTimeProblem — Function

QuantumStateMinimumTimeProblem(traj, sys, obj, integrators, constraints; kwargs...)
QuantumStateMinimumTimeProblem(prob; kwargs...)

Construct a QuantumControlProblem for the minimum time problem of reaching a target state.

Arguments

traj::NamedTrajectory: The initial trajectory.

sys::QuantumSystem: The quantum system.

obj::Objective: The objective function.

integrators::Vector{<:AbstractIntegrator}: The integrators.

constraints::Vector{<:AbstractConstraint}: The constraints.

or

prob::QuantumControlProblem: The quantum control problem.

Keyword Arguments

state_name::Symbol=:ψ̃: The symbol for the state variables.

final_fidelity::Union{Real, Nothing}=nothing: The final fidelity.

D=1.0: The cost weight on the time.

ipopt_options::IpoptOptions=IpoptOptions(): The Ipopt options.

piccolo_options::PiccoloOptions=PiccoloOptions(): The Piccolo options.

kwargs...: Additional keyword arguments, passed to QuantumControlProblem.

QuantumCollocation.ProblemTemplates.QuantumStateSmoothPulseProblem — Function

QuantumStateSmoothPulseProblem(system, ψ_inits, ψ_goals, T, Δt; kwargs...)
QuantumStateSmoothPulseProblem(system, ψ_init, ψ_goal, T, Δt; kwargs...)
QuantumStateSmoothPulseProblem(H_drift, H_drives, args...; kwargs...)

Create a quantum state smooth pulse problem. The goal is to find a control pulse a(t) that drives all of the initial states ψ_inits to the corresponding target states ψ_goals using T timesteps of size Δt. This problem also controls the first and second derivatives of the control pulse, da(t) and dda(t), to ensure smoothness.

Arguments

system::AbstractQuantumSystem: The quantum system.

or

H_drift::AbstractMatrix{<:Number}: The drift Hamiltonian.

H_drives::Vector{<:AbstractMatrix{<:Number}}: The control Hamiltonians.

with

ψ_inits::Vector{<:AbstractVector{<:ComplexF64}}: The initial states.

ψ_goals::Vector{<:AbstractVector{<:ComplexF64}}: The target states.

or

ψ_init::AbstractVector{<:ComplexF64}: The initial state.

ψ_goal::AbstractVector{<:ComplexF64}: The target state.

with

T::Int: The number of timesteps.

Δt::Float64: The timestep size.

Keyword Arguments

ipopt_options::IpoptOptions=IpoptOptions(): The IPOPT options.

piccolo_options::PiccoloOptions=PiccoloOptions(): The Piccolo options.

state_name::Symbol=:ψ̃: The name of the state variable.

control_name::Symbol=:a: The name of the control variable.

timestep_name::Symbol=:Δt: The name of the timestep variable.

init_trajectory::Union{NamedTrajectory, Nothing}=nothing: The initial trajectory.

a_bound::Float64=1.0: The bound on the control pulse.

a_bounds::Vector{Float64}=fill(a_bound, length(system.G_drives)): The bounds on the control pulse.

a_guess::Union{Matrix{Float64}, Nothing}=nothing: The initial guess for the control pulse.

da_bound::Float64=Inf: The bound on the first derivative of the control pulse.

da_bounds::Vector{Float64}=fill(da_bound, length(system.G_drives)): The bounds on the first derivative of the control pulse.

dda_bound::Float64=1.0: The bound on the second derivative of the control pulse.

dda_bounds::Vector{Float64}=fill(dda_bound, length(system.G_drives)): The bounds on the second derivative of the control pulse.

Δt_min::Float64=0.5 * Δt: The minimum timestep size.

Δt_max::Float64=1.5 * Δt: The maximum timestep size.

drive_derivative_σ::Float64=0.01: The standard deviation of the drive derivative random initialization.

Q::Float64=100.0: The weight on the state objective.

R=1e-2: The weight on the control pulse and its derivatives.

R_a::Union{Float64, Vector{Float64}}=R: The weight on the control pulse.

R_da::Union{Float64, Vector{Float64}}=R: The weight on the first derivative of the control pulse.

R_dda::Union{Float64, Vector{Float64}}=R: The weight on the second derivative of the control pulse.

leakage_operator::Union{Nothing, EmbeddedOperator}=nothing: The leakage operator, if leakage suppression is desired.

constraints::Vector{<:AbstractConstraint}=AbstractConstraint[]: The constraints.

QuantumCollocation.ProblemTemplates.UnitaryBangBangProblem — Function

UnitaryBangBangProblem(system::QuantumSystem, operator, T, Δt; kwargs...)
UnitaryBangBangProblem(H_drift, H_drives, operator, T, Δt; kwargs...)

Construct a QuantumControlProblem for a free-time unitary gate problem with bang-bang control pulses.

\[\begin{aligned} \underset{\vec{\tilde{U}}, a, \dot{a}, \Delta t}{\text{minimize}} & \quad Q \cdot \ell\qty(\vec{\tilde{U}}_T, \vec{\tilde{U}}_{\text{goal}}) + R_{\text{bang-bang}} \cdot \sum_t |\dot{a}_t| \\ \text{ subject to } & \quad \vb{P}^{(n)}\qty(\vec{\tilde{U}}_{t+1}, \vec{\tilde{U}}_t, a_t, \Delta t_t) = 0 \\ & \quad a_{t+1} - a_t - \dot{a}_t \Delta t_t = 0 \\ & \quad |a_t| \leq a_{\text{bound}} \\ & \quad |\dot{a}_t| \leq da_{\text{bound}} \\ & \quad \Delta t_{\text{min}} \leq \Delta t_t \leq \Delta t_{\text{max}} \\ \end{aligned}\]

where, for $U \in SU(N)$,

\[\ell\qty(\vec{\tilde{U}}_T, \vec{\tilde{U}}_{\text{goal}}) = \abs{1 - \frac{1}{N} \abs{ \tr \qty(U_{\text{goal}}, U_T)} }\]

is the infidelity objective function, $Q$ is a weight, $R_a$, and $R_{\dot{a}}$ are weights on the regularization terms, and $\vb{P}^{(n)}$ is the $n$th-order Pade integrator.

TODO: Document bang-bang modification.

Arguments

H_drift::AbstractMatrix{<:Number}: the drift hamiltonian

H_drives::Vector{<:AbstractMatrix{<:Number}}: the control hamiltonians

or

system::QuantumSystem: the system to be controlled

with

operator::OperatorType: the target unitary, either in the form of an EmbeddedOperator or a `Matrix{ComplexF64}

T::Int: the number of timesteps

Δt::Float64: the (initial) time step size

Keyword Arguments

ipopt_options::IpoptOptions=IpoptOptions(): the options for the Ipopt solver

piccolo_options::PiccoloOptions=PiccoloOptions(): the options for the Piccolo solver

state_name::Symbol = :Ũ⃗: the name of the state variable

control_name::Symbol = :a: the name of the control variable

timestep_name::Symbol = :Δt: the name of the timestep variable

init_trajectory::Union{NamedTrajectory, Nothing}=nothing: an initial trajectory to use

a_bound::Float64=1.0: the bound on the control pulse

a_bounds=fill(a_bound, length(system.G_drives)): the bounds on the control pulses, one for each drive

a_guess::Union{Matrix{Float64}, Nothing}=nothing: an initial guess for the control pulses

da_bound::Float64=1.0: the bound on the control pulse derivative

da_bounds=fill(da_bound, length(system.G_drives)): the bounds on the control pulse derivatives, one for each drive

Δt_min::Float64=Δt isa Float64 ? 0.5 * Δt : 0.5 * mean(Δt): the minimum time step size

Δt_max::Float64=Δt isa Float64 ? 1.5 * Δt : 1.5 * mean(Δt): the maximum time step size

drive_derivative_σ::Float64=0.01: the standard deviation of the initial guess for the control pulse derivatives

Q::Float64=100.0: the weight on the infidelity objective

R=1e-2: the weight on the regularization terms

quadratic_control_regularization=false: whether or not to use quadratic regularization for the control pulses

R_a::Union{Float64, Vector{Float64}}=R: the weight on the regularization term for the control pulses

R_da::Union{Float64, Vector{Float64}}=R: the weight on the regularization term for the control pulse derivatives

R_bang_bang::Union{Float64, Vector{Float64}}=1e-1: the weight on the bang-bang regularization term

phase_operators::Union{AbstractVector{<:AbstractMatrix}, Nothing}=nothing: the phase operators for free phase corrections

constraints::Vector{<:AbstractConstraint}=AbstractConstraint[]: the constraints to enforce

QuantumCollocation.ProblemTemplates.UnitaryDirectSumProblem — Method

UnitaryDirectSumProblem(probs, final_fidelity; kwargs...)

Construct a QuantumControlProblem as a direct sum of unitary gate problems. The purpose is to find solutions that are as close as possible with respect to one of their components. In particular, this is useful for finding interpolatable control solutions.

A graph of edges (specified by problem labels) will enforce a PairwiseQuadraticRegularizer between the component trajectories of the problem in probs corresponding to the names of the edge in edges with corresponding edge weight Q.

Boundary values can be included to enforce a QuadraticRegularizer on edges where one of the nodes is not optimized. The boundary values are specified as a dictionary with keys corresponding to the edge labels and values corresponding to the boundary values.

The default behavior is to use a 1D chain for the graph, i.e., enforce a PairwiseQuadraticRegularizer between each neighbor of the provided probs.

Arguments

probs::AbstractVector{<:QuantumControlProblem}: the problems to combine

final_fidelity::Real: the fidelity to enforce between the component final unitaries and the component goal unitaries

Keyword Arguments

prob_labels::AbstractVector{<:String}}: the labels for the problems

graph::Union{Nothing, AbstractVector{<:Tuple{String, String}}, AbstractVector{<:Tuple{Symbol, Symbol}}}`: the graph of edges to enforce

boundary_values::Union{Nothing, AbstractDict{<:String, <:AbstractArray}, AbstractDict{<:Symbol, <:AbstractVector}}=nothing: the boundary values for the problems

Q::Union{Float64, Vector{Float64}}=100.0: the weights on the pairwise regularizers

Q_symb::Symbol=:Ũ⃗: the symbol to use for the regularizer

R::Float64=1e-2: the shared weight on all control terms (:a, :da, :dda is assumed)

R_a::Union{Float64, Vector{Float64}}=R: the weight on the regularization term for the control pulses

R_da::Union{Float64, Vector{Float64}}=R: the weight on the regularization term for the control pulse derivatives

R_dda::Union{Float64, Vector{Float64}}=R: the weight on the regularization term for the control pulse second derivatives

R_b::Union{Float64, Vector{Float64}}=R: the weight on the regularization term for the boundary values

drive_derivative_σ::Float64=0.01: the standard deviation of the initial guess for the control pulse derivatives

drive_reset_ratio::Float64=0.1: amount of random noise to add to the control data (can help avoid hitting restoration if provided problems are converged)

fidelity_cost::Bool=false: whether or not to include a fidelity cost in the objective

subspace::Union{AbstractVector{<:Integer}, Nothing}=nothing: the subspace to use for the fidelity of each problem

ipopt_options::IpoptOptions=IpoptOptions(): the options for the Ipopt solver

piccolo_options::PiccoloOptions=PiccoloOptions(): the options for the Piccolo solver

QuantumCollocation.ProblemTemplates.UnitaryRobustnessProblem — Function

UnitaryRobustnessProblem(
    H_error,
    trajectory,
    system,
    objective,
    integrators,
    constraints;
    kwargs...
)

UnitaryRobustnessProblem(Hₑ, prob::QuantumControlProblem; kwargs...)

Create a quantum control problem for robustness optimization of a unitary trajectory.

Keyword Arguments

unitary_name::Symbol=:Ũ⃗: The symbol for the unitary trajectory in trajectory.

final_fidelity::Union{Real, Nothing}=nothing: The target fidelity for the final unitary.

ipopt_options::IpoptOptions=IpoptOptions(): Options for the Ipopt solver.

piccolo_options::PiccoloOptions=PiccoloOptions(): Options for the Piccolo solver.

kwargs...: Additional keyword arguments passed to QuantumControlProblem.

QuantumCollocation.ProblemTemplates.UnitarySamplingProblem — Method

UnitarySamplingProblem(systemns, operator, T, Δt; kwargs...)

A UnitarySamplingProblem is a quantum control problem where the goal is to find a control pulse that generates a target unitary operator for a set of quantum systems. The controls are shared among all systems, and the optimization seeks to find the control pulse that achieves the operator for each system. The idea is to enforce a robust solution by including multiple systems reflecting the problem uncertainty.

Arguments

systems::AbstractVector{<:AbstractQuantumSystem}: A vector of quantum systems.

operator::OperatorType: The target unitary operator.

T::Int: The number of time steps.

Δt::Union{Float64, Vector{Float64}}: The time step value or vector of time steps.

Keyword Arguments

system_labels::Vector{String} = string.(1:length(systems)): The labels for each system.

system_weights::Vector{Float64} = fill(1.0, length(systems)): The weights for each system.

init_trajectory::Union{NamedTrajectory, Nothing} = nothing: The initial trajectory.

ipopt_options::IpoptOptions = IpoptOptions(): The IPOPT options.

piccolo_options::PiccoloOptions = PiccoloOptions(): The Piccolo options.

state_name::Symbol = :Ũ⃗: The name of the state variable.

control_name::Symbol = :a: The name of the control variable.

timestep_name::Symbol = :Δt: The name of the timestep variable.

constraints::Vector{<:AbstractConstraint} = AbstractConstraint[]: The constraints.

a_bound::Float64 = 1.0: The bound for the control amplitudes.

a_bounds::Vector{Float64} = fill(a_bound, length(systems[1].G_drives)): The bounds for the control amplitudes.

a_guess::Union{Matrix{Float64}, Nothing} = nothing: The initial guess for the control amplitudes.

da_bound::Float64 = Inf: The bound for the control first derivatives.

da_bounds::Vector{Float64} = fill(da_bound, length(systems[1].G_drives)): The bounds for the control first derivatives.

dda_bound::Float64 = 1.0: The bound for the control second derivatives.

dda_bounds::Vector{Float64} = fill(dda_bound, length(systems[1].G_drives)): The bounds for the control second derivatives.

Δt_min::Float64 = 0.5 * Δt: The minimum time step size.

Δt_max::Float64 = 1.5 * Δt: The maximum time step size.

Q::Float64 = 100.0: The fidelity weight.

R::Float64 = 1e-2: The regularization weight.

R_a::Union{Float64, Vector{Float64}} = R: The regularization weight for the control amplitudes.

R_da::Union{Float64, Vector{Float64}} = R: The regularization weight for the control first derivatives.

R_dda::Union{Float64, Vector{Float64}} = R: The regularization weight for the control second derivatives.

kwargs...: Additional keyword arguments.

QuantumCollocation.ProblemTemplates.UnitarySmoothPulseProblem — Method

UnitarySmoothPulseProblem(system::AbstractQuantumSystem, operator, T, Δt; kwargs...)
UnitarySmoothPulseProblem(H_drift, H_drives, operator, T, Δt; kwargs...)

Construct a QuantumControlProblem for a free-time unitary gate problem with smooth control pulses enforced by constraining the second derivative of the pulse trajectory, i.e.,

\[\begin{aligned} \underset{\vec{\tilde{U}}, a, \dot{a}, \ddot{a}, \Delta t}{\text{minimize}} & \quad Q \cdot \ell\qty(\vec{\tilde{U}}_T, \vec{\tilde{U}}_{\text{goal}}) + \frac{1}{2} \sum_t \qty(R_a a_t^2 + R_{\dot{a}} \dot{a}_t^2 + R_{\ddot{a}} \ddot{a}_t^2) \\ \text{ subject to } & \quad \vb{P}^{(n)}\qty(\vec{\tilde{U}}_{t+1}, \vec{\tilde{U}}_t, a_t, \Delta t_t) = 0 \\ & \quad a_{t+1} - a_t - \dot{a}_t \Delta t_t = 0 \\ & \quad \dot{a}_{t+1} - \dot{a}_t - \ddot{a}_t \Delta t_t = 0 \\ & \quad |a_t| \leq a_{\text{bound}} \\ & \quad |\ddot{a}_t| \leq \ddot{a}_{\text{bound}} \\ & \quad \Delta t_{\text{min}} \leq \Delta t_t \leq \Delta t_{\text{max}} \\ \end{aligned}\]

where, for $U \in SU(N)$,

\[\ell\qty(\vec{\tilde{U}}_T, \vec{\tilde{U}}_{\text{goal}}) = \abs{1 - \frac{1}{N} \abs{ \tr \qty(U_{\text{goal}}, U_T)} }\]

is the infidelity objective function, $Q$ is a weight, $R_a$, $R_{\dot{a}}$, and $R_{\ddot{a}}$ are weights on the regularization terms, and $\vb{P}^{(n)}$ is the $n$th-order Pade integrator.

Arguments

system::AbstractQuantumSystem: the system to be controlled

or

H_drift::AbstractMatrix{<:Number}: the drift hamiltonian

H_drives::Vector{<:AbstractMatrix{<:Number}}: the control hamiltonians

with

operator::OperatorType: the target unitary, either in the form of an EmbeddedOperator or a `Matrix{ComplexF64}

T::Int: the number of timesteps

Δt::Float64: the (initial) time step size

Keyword Arguments

ipopt_options::IpoptOptions=IpoptOptions(): the options for the Ipopt solver

piccolo_options::PiccoloOptions=PiccoloOptions(): the options for the Piccolo solver

state_name::Symbol = :Ũ⃗: the name of the state

control_name::Symbol = :a: the name of the control

timestep_name::Symbol = :Δt: the name of the timestep

init_trajectory::Union{NamedTrajectory, Nothing}=nothing: an initial trajectory to use

a_bound::Float64=1.0: the bound on the control pulse

a_bounds::Vector{Float64}=fill(a_bound, length(system.G_drives)): the bounds on the control pulses, one for each drive

a_guess::Union{Matrix{Float64}, Nothing}=nothing: an initial guess for the control pulses

da_bound::Float64=Inf: the bound on the control pulse derivative

da_bounds::Vector{Float64}=fill(da_bound, length(system.G_drives)): the bounds on the control pulse derivatives, one for each drive

dda_bound::Float64=1.0: the bound on the control pulse second derivative

dda_bounds::Vector{Float64}=fill(dda_bound, length(system.G_drives)): the bounds on the control pulse second derivatives, one for each drive

Δt_min::Float64=Δt isa Float64 ? 0.5 * Δt : 0.5 * mean(Δt): the minimum time step size

Δt_max::Float64=Δt isa Float64 ? 1.5 * Δt : 1.5 * mean(Δt): the maximum time step size

Q::Float64=100.0: the weight on the infidelity objective

R=1e-2: the weight on the regularization terms

R_a::Union{Float64, Vector{Float64}}=R: the weight on the regularization term for the control pulses

R_da::Union{Float64, Vector{Float64}}=R: the weight on the regularization term for the control pulse derivatives

R_dda::Union{Float64, Vector{Float64}}=R: the weight on the regularization term for the control pulse second derivatives

phase_name::Symbol=:ϕ: the name of the phase

phase_operators::Union{AbstractVector{<:AbstractMatrix}, Nothing}=nothing: the phase operators for free phase corrections

constraints::Vector{<:AbstractConstraint}=AbstractConstraint[]: the constraints to enforce

Direct Sums

QuantumCollocation.DirectSums.direct_sum — Method

direct_sum(A::AbstractMatrix, B::AbstractMatrix)

Returns the direct sum of two matrices.

QuantumCollocation.DirectSums.direct_sum — Method

direct_sum(Ã⃗::AbstractVector, B̃⃗::AbstractVector)

Returns the direct sum of two iso_vec operators.

QuantumCollocation.DirectSums.direct_sum — Method

direct_sum(traj1::NamedTrajectory, traj2::NamedTrajectory)

Returns the direct sum of two NamedTrajectory objects.

The NamedTrajectory objects must have the same timestep. However, a direct sum can return a free time problem by passing the keyword argument free_time=true. In this case, the timestep symbol must be provided. If a free time problem with more than two trajectories is desired, the reduce function has been written to handle calls to direct sums of NamedTrajectory objects; simply pass the keyword argument free_time=true to the reduce function.

Arguments

traj1::NamedTrajectory: The first NamedTrajectory object.

traj2::NamedTrajectory: The second NamedTrajectory object.

free_time::Bool=false: Whether to construct a free time problem.

timestep_name::Symbol=:Δt: The timestep symbol to use for free time problems.

QuantumCollocation.DirectSums.direct_sum — Method

direct_sum(sys1::QuantumSystem, sys2::QuantumSystem)

Returns the direct sum of two QuantumSystem objects.

QuantumCollocation.DirectSums.direct_sum — Method

direct_sum(A::SparseMatrixCSC, B::SparseMatrixCSC)

Returns the direct sum of two sparse matrices.

Quantum Object Utils

QuantumCollocation.QuantumObjectUtils.GATES — Constant

A constant dictionary GATES containing common quantum gate matrices as complex-valued matrices. Each gate is represented by its unitary matrix.

GATES[:I] - Identity gate: Leaves the state unchanged.

GATES[:X] - Pauli-X (NOT) gate: Flips the qubit state.

GATES[:Y] - Pauli-Y gate: Rotates the qubit state around the Y-axis of the Bloch sphere.

GATES[:Z] - Pauli-Z gate: Flips the phase of the qubit state.

GATES[:H] - Hadamard gate: Creates superposition by transforming basis states.

GATES[:CX] - Controlled-X (CNOT) gate: Flips the second qubit (target) if the first qubit (control) is |1⟩.

GATES[:CZ] - Controlled-Z (CZ) gate: Flips the phase of the second qubit (target) if the first qubit (control) is |1⟩.

GATES[:XI] - Complex gate: A specific gate used for complex operations.

GATES[:sqrtiSWAP] - Square root of iSWAP gate: Partially swaps two qubits with a phase.

julia> GATES[:Z]
2×2 Matrix{ComplexF64}:
 1.0+0.0im   0.0+0.0im
 0.0+0.0im  -1.0+0.0im

julia> get_gate(:CX)
4×4 Matrix{ComplexF64}:
 1.0+0.0im  0.0+0.0im  0.0+0.0im  0.0+0.0im
 0.0+0.0im  1.0+0.0im  0.0+0.0im  0.0+0.0im
 0.0+0.0im  0.0+0.0im  0.0+0.0im  1.0+0.0im
 0.0+0.0im  0.0+0.0im  1.0+0.0im  0.0+0.0im

QuantumCollocation.QuantumObjectUtils.:⊗ — Method

⊗(A::AbstractVecOrMat, B::AbstractVecOrMat) = kron(A, B)

The Kronecker product, denoted by ⊗, results in a block matrix formed by multiplying each element of A by the entire matrix B.

julia> GATES[:X] ⊗ GATES[:Y]
4×4 Matrix{ComplexF64}:
 0.0+0.0im  0.0+0.0im  0.0+0.0im  0.0-1.0im
 0.0+0.0im  0.0+0.0im  0.0+1.0im  0.0+0.0im
 0.0+0.0im  0.0-1.0im  0.0+0.0im  0.0+0.0im
 0.0+1.0im  0.0+0.0im  0.0+0.0im  0.0+0.0im

QuantumCollocation.QuantumObjectUtils.annihilate — Method

annihilate(levels::Int)

Get the annihilation operator for a system with levels levels.

QuantumCollocation.QuantumObjectUtils.create — Method

create(levels::Int)

Get the creation operator for a system with levels levels.

QuantumCollocation.QuantumObjectUtils.haar_identity — Method

haar_identity(n::Int, radius::Number)

Generate a random unitary matrix close to the identity matrix using the Haar measure for an n-dimensional system with a given radius.

QuantumCollocation.QuantumObjectUtils.haar_random — Method

haar_random(n::Int)

Generate a random unitary matrix using the Haar measure for an n-dimensional system.

QuantumCollocation.QuantumObjectUtils.ket_from_bitstring — Method

ket_from_bitstring(ket::String)

Get the state vector for a qubit system given a ket string ket of 0s and 1s.

QuantumCollocation.QuantumObjectUtils.ket_from_string — Method

ket_from_string(
    ket::String,
    levels::Vector{Int};
    level_dict=Dict(:g => 0, :e => 1, :f => 2, :h => 2),
    return_states=false
)

Construct a quantum state from a string ket representation.

Example

TODO: add example

QuantumCollocation.QuantumObjectUtils.operator_from_string — Method

operatorfromstring(operator::String; lookup::Dict{Symbol, AbstractMatrix}=PAULIS)

Reduce the string (each character is one key) via operators from a dictionary.

Quantum Systems

QuantumCollocation.QuantumSystems.AbstractQuantumSystem — Type

AbstractQuantumSystem

Abstract type for defining systems.

QuantumCollocation.QuantumSystems.QuantumSystem — Type

QuantumSystem <: AbstractQuantumSystem

A struct for storing the isomorphisms of the system's drift and drive Hamiltonians, as well as the system's parameters.

QuantumCollocation.QuantumSystems.QuantumSystem — Method

QuantumSystem(
    H_drift::Matrix{<:Number},
    H_drives::Vector{Matrix{<:Number}};
    params=Dict{Symbol, Any}(),
    kwargs...
)::QuantumSystem

Constructs a QuantumSystem object from the drift and drive Hamiltonian terms.

QuantumCollocation.QuantumSystems.QuantumSystemCoupling — Type

QuantumSystemCoupling <: AbstractQuantumSystem

QuantumCollocation.QuantumSystems.lift — Method

lift(U::AbstractMatrix{<:Number}, qubit_index::Int, n_qubits::Int; levels::Int=size(U, 1))

Lift an operator U acting on a single qubit to an operator acting on the entire system of n_qubits.

QuantumCollocation.QuantumSystems.lift — Method

lift(op::AbstractMatrix{<:Number}, i::Int, subsystem_levels::Vector{Int})

Lift an operator op acting on the i-th subsystem to an operator acting on the entire system with given subsystem levels.

Integrators

QuantumCollocation.Integrators.UnitaryPadeIntegrator — Type

Objectives

QuantumCollocation.Objectives.Objective — Type

Objective

A structure for defining objective functions.

The terms field contains all the arguments needed to construct the objective function.

Fields: L: the objective function ∇L: the gradient of the objective function ∂²L: the Hessian of the objective function ∂²L_structure: the structure of the Hessian of the objective function terms: a vector of dictionaries containing the terms of the objective function

QuantumCollocation.Objectives.L1Regularizer — Method

L1Regularizer

Create an L1 regularizer for the trajectory component. The regularizer is defined as

\[J_{L1}(u) = \sum_t \abs{R \cdot u_t}\]

where (R) is the regularization matrix and (u_t) is the trajectory component at time (t).

QuantumCollocation.Objectives.MinimumTimeObjective — Method

MinimumTimeObjective

A type of objective that counts the time taken to complete a task.

Fields: D: a scaling factor Δt_indices: the indices of the time steps eval_hessian: whether to evaluate the Hessian

QuantumCollocation.Objectives.PairwiseQuadraticRegularizer — Method

PairwiseQuadraticRegularizer

Create a pairwise quadratic regularizer for the trajectory component name with regularization strength R. The regularizer is defined as

\[ J_{v⃗}(u) = \sum_t \frac{1}{2} \Delta t_t^2 (v⃗_{1,t} - v⃗_{2,t})^T R (v⃗_{1,t} - v⃗_{2,t})\]

where $v⃗_{1}$ and $v⃗_{2}$ are selected by name1 and name2. The indices specify the appropriate block diagonal components of the direct sum vector v⃗.

TODO: Hessian not implemented

Fields: R: the regularization strength times: the time steps to apply the regularizer name1: the first name name2: the second name timestep_name: the symbol for the timestep eval_hessian: whether to evaluate the Hessian

QuantumCollocation.Objectives.PairwiseQuadraticRegularizer — Method

PairwiseQuadraticRegularizer

A convenience constructor for creating a PairwiseQuadraticRegularizer for the trajectory component name with regularization strength Rs over the graph graph.

QuantumCollocation.Objectives.PairwiseUnitaryRobustnessObjective — Method

PairwiseUnitaryRobustnessObjective(;
    H1::Union{OperatorType, Nothing}=nothing,
    H2_error::Union{OperatorType, Nothing}=nothing,
    symb1::Symbol=:Ũ⃗1,
    symb2::Symbol=:Ũ⃗2,
    eval_hessian::Bool=false,
)

Create a control objective which penalizes the sensitivity of the infidelity to the provided operators defined in the subspaces of the control dynamics, thereby realizing robust control.

QuantumCollocation.Objectives.QuadraticRegularizer — Method

QuadraticRegularizer

A quadratic regularizer for a trajectory component.

Fields: name: the name of the trajectory component to regularize times: the times at which to evaluate the regularizer dim: the dimension of the trajectory component R: the regularization matrix baseline: the baseline values for the trajectory component eval_hessian: whether to evaluate the Hessian of the regularizer timestep_name: the symbol for the timestep variable

QuantumCollocation.Objectives.QuadraticSmoothnessRegularizer — Method

QuadraticSmoothnessRegularizer

A quadratic smoothness regularizer for a trajectory component.

Fields: name: the name of the trajectory component to regularize times: the times at which to evaluate the regularizer R: the regularization matrix eval_hessian: whether to evaluate the Hessian of the regularizer

QuantumCollocation.Objectives.QuantumObjective — Method

QuantumObjective

A generic objective function for quantum trajectories that use a loss.

QuantumCollocation.Objectives.UnitaryFreePhaseInfidelityObjective — Method

UnitaryFreePhaseInfidelityObjective

A type of objective that measures the infidelity of a unitary operator to a target unitary operator, where the target unitary operator is allowed to have phases on qubit subspaces.

Fields: name: the name of the unitary operator in the trajectory global_name: the name of the global phase in the trajectory goal: the target unitary operator Q: a scaling factor eval_hessian: whether to evaluate the Hessian subspace: the subspace in which to evaluate the objective

QuantumCollocation.Objectives.UnitaryInfidelityObjective — Method

UnitaryInfidelityObjective

A type of objective that measures the infidelity of a unitary operator to a target unitary operator.

Fields: name: the name of the unitary operator in the trajectory goal: the target unitary operator Q: a scaling factor eval_hessian: whether to evaluate the Hessian subspace: the subspace in which to evaluate the objective

QuantumCollocation.Objectives.UnitaryRobustnessObjective — Method

UnitaryRobustnessObjective(; H::::Union{OperatorType, Nothing}=nothing, eval_hessian::Bool=false, symb::Symbol=:Ũ⃗ )

Create a control objective which penalizes the sensitivity of the infidelity to the provided operator defined in the subspace of the control dynamics, thereby realizing robust control.

The control dynamics are

\[U_C(a)= \prod_t \exp{-i H_C(a_t)}\]

In the control frame, the H operator is (proportional to)

\[R_{Robust}(a) = \frac{1}{T \norm{H_e}_2} \sum_t U_C(a_t)^\dag H_e U_C(a_t) \Delta t\]

where we have adjusted to a unitless expression of the operator.

The robustness objective is

\[R_{Robust}(a) = \frac{1}{N} \norm{R}^2_F\]

where N is the dimension of the Hilbert space.

QuantumCollocation.Objectives.sparse_to_moi — Method

sparse_to_moi(A::SparseMatrixCSC)

Converts a sparse matrix to tuple of vector of nonzero indices and vector of nonzero values

Losses

QuantumCollocation.Losses.fidelity — Method

fidelity(ψ, ψ_goal)

Calculate the fidelity between two quantum states ψ and ψ_goal.

QuantumCollocation.Losses.iso_fidelity — Method

iso_fidelity(ψ̃, ψ̃_goal)

Calculate the fidelity between two quantum states in their isomorphic form ψ̃ and ψ̃_goal.

QuantumCollocation.Losses.iso_infidelity — Function

iso_infidelity(ψ̃, ψ̃goal)

Returns the iso_infidelity between two quantum statevectors specified in the $\mathbb{C}^n \to \mathbb{R}^{2n}$ isomorphism space.

QuantumCollocation.Losses.iso_vec_unitary_fidelity — Method

iso_vec_unitary_fidelity(Ũ⃗::AbstractVector, Ũ⃗_goal::AbstractVector)

Returns the fidelity between the isomorphic unitary vector $\vec{\widetilde{U}} \sim U \in SU(n)$ and the isomorphic goal unitary vector $\vec{\widetilde{U}}_{\text{goal}}$.

\[\begin{aligned} \mathcal{F}(\vec{\widetilde{U}}, \vec{\widetilde{U}}_{\text{goal}}) &= \frac{1}{n} \abs{\tr \qty(U_{\text{goal}}^\dagger U)} \\ &= \frac{1}{n} \sqrt{T_R^{2} + T_I^{2}} \end{aligned}\]

where $T_R = \langle \vec{\widetilde{U}}_{\text{goal}, R}, \vec{\widetilde{U}}_R \rangle + \langle \vec{\widetilde{U}}_{\text{goal}, I}, \vec{\widetilde{U}}_I \rangle$ and $T_I = \langle \vec{\widetilde{U}}_{\text{goal}, R}, \vec{\widetilde{U}}_I \rangle - \langle \vec{\widetilde{U}}_{\text{goal}, I}, \vec{\widetilde{U}}_R \rangle$.

QuantumCollocation.Losses.unitary_fidelity — Method

unitary_fidelity(U::Matrix, U_goal::Matrix; kwargs...)
unitary_fidelity(Ũ⃗::AbstractVector, Ũ⃗_goal::AbstractVector; kwargs...)

Calculate the fidelity between two unitary operators U and U_goal.

\[\mathcal{F}(U, U_{\text{goal}}) = \frac{1}{n} \abs{\tr \qty(U_{\text{goal}}^\dagger U)}\]

where $n$ is the dimension of the unitary operators.

Keyword Arguments

subspace::AbstractVector{Int}: The subspace to calculate the fidelity over.

Embedded Operators

QuantumCollocation.EmbeddedOperators.EmbeddedOperator — Method

EmbeddedOperator(op::Matrix{<:Number}, subspace_indices::AbstractVector{Int}, subsystem_levels::AbstractVector{Int})

Create an embedded operator. The operator op is embedded in the subspace defined by subspace_indices in subsystem_levels.

Arguments

op::Matrix{<:Number}: Operator to embed.

subspace_indices::AbstractVector{Int}: Indices of the subspace to embed the operator in. e.g. get_subspace_indices([1:2, 1:2], [3, 3]).

subsystem_levels::AbstractVector{Int}: Levels of the subsystems in the composite system. e.g. [3, 3] for two 3-level systems.

QuantumCollocation.EmbeddedOperators.embed — Method

embed(matrix::Matrix{ComplexF64}, subspace_indices::AbstractVector{Int}, levels::Int)

Embed an operator $U$ in the subspace of a larger system $\mathcal{X} = \mathcal{X}_{\text{subspace}} \oplus \mathcal{X}_{\text{leakage}}$ which is composed of matrices of size $\text{levels} \times \text{levels}$.

Arguments

matrix::Matrix{ComplexF64}: Operator to embed.

subspace_indices::AbstractVector{Int}: Indices of the subspace to embed the operator in.

levels::Int: Total number of levels in the system.

QuantumCollocation.EmbeddedOperators.get_subspace_indices — Method

get_subspace_indices(subspaces::Vector{<:AbstractVector{Int}}, subsystem_levels::AbstractVector{Int})

Get the indices for the subspace of composite quantum system.

Example: for the two-qubit subspace of two 3-level systems:

subspaces = [1:2, 1:2]
subsystem_levels = [3, 3]
get_subspace_indices(subspaces, subsystem_levels) == [1, 2, 4, 5]

Arguments

subspaces::Vector{<:AbstractVector{Int}}: Subspaces to get indices for. e.g. [1:2, 1:2].

subsystem_levels::AbstractVector{Int}: Levels of the subsystems in the composite system. e.g. [3, 3]. Each element corresponds to a subsystem.

QuantumCollocation.EmbeddedOperators.unembed — Method

unembed(matrix::AbstractMatrix, subspace_indices::AbstractVector{Int})

Unembed an operator $U$ from a subspace of a larger system $\mathcal{X} = \mathcal{X}_{\text{subspace}} \oplus \mathcal{X}_{\text{leakage}}$ which is composed of matrices of size $\text{levels} \times \text{levels}$.

This is equivalent to calling matrix[subspace_indices, subspace_indices].

Arguments

matrix::AbstractMatrix: Operator to unembed.

subspace_indices::AbstractVector{Int}: Indices of the subspace to unembed the operator from.

Isomorphisms

QuantumCollocation.Isomorphisms.G — Method

G(H::AbstractMatrix)::Matrix{Float64}

Returns the isomorphism of $-iH$:

\[G(H) = \widetilde{- i H} = \mqty(1 & 0 \\ 0 & 1) \otimes \Im(H) - \mqty(0 & -1 \\ 1 & 0) \otimes \Re(H)\]

where $\Im(H)$ and $\Re(H)$ are the imaginary and real parts of $H$ and the tilde indicates the standard isomorphism of a complex valued matrix:

\[\widetilde{H} = \mqty(1 & 0 \\ 0 & 1) \otimes \Re(H) + \mqty(0 & -1 \\ 1 & 0) \otimes \Im(H)\]

QuantumCollocation.Isomorphisms.H — Method

H(G::AbstractMatrix{<:Number})::Matrix{ComplexF64}

Returns the inverse of G(H) = iso(-iH), i.e. returns H

QuantumCollocation.Isomorphisms.iso_dm — Method

iso_dm(ρ::AbstractMatrix)

returns the isomorphism ρ⃗̃ = ket_to_iso(vec(ρ)) of a density matrix ρ

QuantumCollocation.Isomorphisms.iso_operator_to_iso_vec — Method

iso_operator_to_iso_vec(Ũ::AbstractMatrix)

Convert a real matrix Ũ representing an isomorphism operator into a real vector.

Must be differentiable.

QuantumCollocation.Isomorphisms.iso_to_ket — Method

iso_to_ket(ψ̃)

Convert a complex vector ψ̃ with real and imaginary parts into a ket vector.

QuantumCollocation.Isomorphisms.iso_vec_to_iso_operator — Method

iso_vec_to_iso_operator(Ũ⃗::AbstractVector)

Convert a real vector Ũ⃗ into a real matrix representing an isomorphism operator.

Must be differentiable.

QuantumCollocation.Isomorphisms.iso_vec_to_operator — Method

iso_vec_to_operator(Ũ⃗::AbstractVector)

Convert a real vector Ũ⃗ into a complex matrix representing an operator.

Must be differentiable.

QuantumCollocation.Isomorphisms.ket_to_iso — Method

ket_to_iso(ψ)

Convert a ket vector ψ into a complex vector with real and imaginary parts.

QuantumCollocation.Isomorphisms.mat — Method

mat(x::AbstractVector)

Convert a vector x into a square matrix. The length of x must be a perfect square.

QuantumCollocation.Isomorphisms.operator_to_iso_vec — Method

operator_to_iso_vec(U::AbstractMatrix{<:Complex})

Convert a complex matrix U representing an operator into a real vector.

Must be differentiable.

Options

QuantumCollocation.Options.IpoptOptions — Type

Solver options for Ipopt

https://coin-or.github.io/Ipopt/OPTIONS.html#OPT_print_options_documentation

QuantumCollocation.Options.PiccoloOptions — Type

PiccoloOptions

Options for the Piccolo quantum optimal control library.

Fields

verbose::Bool = true: Print verbose output

verbose_evaluator::Bool = false: Print verbose output from the evaluator

free_time::Bool = true: Allow free time optimization

timesteps_all_equal::Bool = true: Use equal timesteps

integrator::Symbol = :pade: Integrator to use

pade_order::Int = 4: Order of the Pade approximation

rollout_integrator::Function = expv: Integrator to use for rollout

jacobian_structure::Bool = integrator == :pade: Use the integrator's Jacobian structure

eval_hessian::Bool = false: Evaluate the Hessian

geodesic = true: Use the geodesic to initialize the optimization.

blas_multithreading::Bool = true: Use BLAS multithreading.

build_trajectory_constraints::Bool = true: Build trajectory constraints.

complex_control_norm_constraint_name::Union{Nothing, Symbol} = nothing: Name of the complex control norm constraint.

complex_control_norm_constraint_radius::Float64 = 1.0: Radius of the complex control norm constraint.

bound_state::Bool = false: Bound the state.

leakage_suppression::Bool = false: Suppress leakage.

R_leakage::Float64 = 1.0: Leakage suppression parameter.

free_phase_infidelity::Bool = false: Free phase infidelity.

phase_operators::Union{Nothing, AbstractVector{<:AbstractMatrix{<:Complex}}} = nothing: Phase operators.

phase_name::Symbol = :ϕ: Name of the phase.

Plotting

QuantumCollocation.Plotting.plot_unitary_populations — Function

plot_unitary_populations(
    traj::NamedTrajectory;
    unitary_columns::AbstractVector{Int}=1:2,
    unitary_name::Symbol=:Ũ⃗,
    control_name::Symbol=:a,
    kwargs...
)

plot_unitary_populations(
    prob::QuantumControlProblem;
    kwargs...
)

Plot the populations of the unitary columns of the unitary matrix in the trajectory. kwargs are passed to NamedTrajectories.plot.

Problem Solvers

Rollouts

QuantumCollocation.Losses.iso_vec_unitary_fidelity — Method

Compute the rollout fidelity.

QuantumCollocation.Rollouts.infer_is_evp — Method

infer_is_evp(integrator::Function)

Infer whether the integrator is a exponential-vector product (EVP) function.

If true, the integrator is expected to have a signature like the exponential action, expv. Otherwise, it is expected to have a signature like exp.

QuantumCollocation.Rollouts.lab_frame_unitary_rollout — Method

lab_frame_unitary_rollout(
    sys::AbstractQuantumSystem,
    controls::AbstractMatrix{Float64};
    duration=nothing,
    timestep=nothing,
    ω=nothing,
    timestep_nyquist=1 / (50 * ω)
)

QuantumCollocation.Rollouts.rollout — Method

rollout(
    ψ̃_init::AbstractVector{<:Float64},
    controls::AbstractMatrix,
    Δt::AbstractVector,
    system::AbstractQuantumSystem
)

Rollout a quantum state ψ̃_init under the control controls for a time Δt using the system system.

If exp_vector_product is true, the integrator is expected to have a signature like the exponential action, expv. Otherwise, it is expected to have a signature like exp.

Types should allow for autodifferentiable controls and times.

Saving and Loading

Structure Utils

Trajectory Initialization

QuantumCollocation.TrajectoryInitialization.initialize_quantum_states — Method

initialize_quantum_states

Initialize the quantum states for a single set of parameters.

QuantumCollocation.TrajectoryInitialization.initialize_trajectory — Method

initialize_trajectory

Trajectory initialization of quantum states can broadcast over multiple systems.

QuantumCollocation.TrajectoryInitialization.initialize_trajectory — Method

initialize_trajectory

Trajectory initialization of unitary states can broadcast over multiple systems.

QuantumCollocation.TrajectoryInitialization.initialize_trajectory — Method

initialize_trajectory

Initialize a trajectory for a control problem. The trajectory is initialized with data that should be consistently the same type (in this case, Float64).

QuantumCollocation.TrajectoryInitialization.initialize_unitaries — Method

Initialize the unitary states for a single set of parameters.

QuantumCollocation.TrajectoryInitialization.unitary_geodesic — Function

unitary_geodesic(
    operator::EmbeddedOperator,
    samples::Int;
    kwargs...
)

unitary_geodesic(
    U_goal::AbstractMatrix{<:Number},
    samples::Int;
    kwargs...
)

unitary_geodesic(
    U₀::AbstractMatrix{<:Number},
    U₁::AbstractMatrix{<:Number},
    samples::Number;
    kwargs...
)

unitary_geodesic(
    U₀::AbstractMatrix{<:Number},
    U₁::AbstractMatrix{<:Number},
    timesteps::AbstractVector{<:Number};
    return_generator=false
)

Compute a geodesic connecting two unitary operators.

QuantumCollocation.TrajectoryInitialization.unitary_geodesic — Method

unitary_geodesic(U_init, U_goal, times; kwargs...)

Compute the geodesic connecting Uinit and Ugoal at the specified times. Allows for the possibility of unequal times and ranges outside [0,1].

Arguments

U_init::AbstractMatrix{<:Number}: The initial unitary operator.

U_goal::AbstractMatrix{<:Number}: The goal unitary operator.

times::AbstractVector{<:Number}: The times at which to evaluate the geodesic.

Keyword Arguments

return_unitary_isos::Bool=true: If true returns a matrix where each column is a unitary isovec, i.e. vec(vcat(real(U), imag(U))). If false, returns a vector of unitary matrices.

return_generator::Bool=false: If true, returns the effective Hamiltonian generating the geodesic.

QuantumCollocation.TrajectoryInitialization.unitary_linear_interpolation — Method

unitary_linear_interpolation(
    U_init::AbstractMatrix,
    U_goal::AbstractMatrix,
    samples::Int
)

Compute a linear interpolation of unitary operators with samples samples.

Trajectory Interpolations